pyproject.toml
[tool.poetry]
name = "data_preparation"
version = "0.1.0"
description = ""
authors = ["Pattarapon Kittisrisawai <pattarapon.kitti@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
pyspark = "3.2.1"
SQLAlchemy = "1.4.32"
pyodbc = "4.0.32"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
python-dotenv = "^0.19.2"
psycopg2 = "2.9.3"
pandas = "^1.4.1"
openpyxl = "^3.0.9"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.32"}
mypy = "^0.942"
pylint = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120